+++ src/ctest.ts
@@ -241,14 +241,17 @@ export class CTestDriver implements vscode.Disposable {
         return items;
     };
 
-    private async getCTestArgs(driver: CMakeDriver, useCMakePresets: boolean, customizedTask: boolean = false, testPreset?: TestPreset): Promise<string[]> {
+    private async getCTestArgs(driver: CMakeDriver, useCMakePresets: boolean, customizedTask: boolean = false, testPreset?: TestPreset): Promise<string[] | undefined> {
         let ctestArgs: string[];
         if (customizedTask && testPreset) {
             ctestArgs = ['-T', 'test'].concat(testArgs(testPreset));
         } else if (!customizedTask && useCMakePresets) {
-            // The drv.useCMakePresets is not updated here yet.
             if (!driver.testPreset) {
-                throw(localize('test.preset.not.set', 'Test preset is not set'));
+                await vscode.commands.executeCommand('cmake.selectTestPreset', (await this.projectController?.getProjectForFolder(driver.workspaceFolder))?.workspaceFolder);
+            }
+            if (!driver.testPreset) {
+                // Test explorer doesn't handle errors well, so we need to deal with them ourselves
+                return undefined;
             }
             // Add a few more args so we can show the result in status bar
             ctestArgs = ['-T', 'test'].concat(testArgs(driver.testPreset));
@@ -269,6 +272,7 @@ export class CTestDriver implements vscode.Disposable {
         return ctestArgs;
     }
 
+    // The drv.useCMakePresets may not be updated here yet, so we need to pass it in.
     public async runCTest(driver: CMakeDriver, useCMakePresets: boolean, customizedTask: boolean = false, testPreset?: TestPreset, consumer?: proc.OutputConsumer): Promise<number> {
         if (!customizedTask) {
             // We don't want to focus on log channel when running tasks.
@@ -371,6 +375,11 @@ export class CTestDriver implements vscode.Disposable {
                 _ctestArgs = await this.getCTestArgs(_driver, useCMakePresets, customizedTask);
             }
 
+            if (!_ctestArgs) {
+                this.ctestErrored(test, run, { message: localize('ctest.args.not.found', 'Could not get test arguments') });
+                continue;
+            }
+
             if (test.children.size > 0) {
                 // Shouldn't reach here now, but not hard to write so keeping it in case we want to have more complicated test hierarchies
                 const children = this.testItemCollectionToArray(test.children);
@@ -382,51 +391,66 @@ export class CTestDriver implements vscode.Disposable {
 
                 const testResults = await this.runCTestImpl(_driver, _ctestPath, _ctestArgs, test.id, customizedTask, consumer);
 
+                let foundTestResult = false;
+                // Only show the first failure
+                let havefailures = false;
+                let duration: number | undefined;
                 if (testResults) {
-                    if (testResults.site.testing.test.length === 1) {
-                        let output = testResults.site.testing.test[0].output;
+                    for (let i = 0; i < testResults.site.testing.test.length; i++) {
+                        const testName = testResults.site.testing.test[i].name;
+                        if (testName === test.id) {
+                            foundTestResult = true;
+                            const durationStr = testResults.site.testing.test[i].measurements.get("Execution Time")?.value;
+                            duration = durationStr ? parseFloat(durationStr) * 1000 : undefined;
+                        }
+
+                        let output = testResults.site.testing.test[i].output;
                         if (process.platform === 'win32') {
                             output = output.replace(/\r?\n/g, '\r\n');
                         }
                         run.appendOutput(output);
 
-                        const durationStr = testResults.site.testing.test[0].measurements.get("Execution Time")?.value;
-                        const duration = durationStr ? parseFloat(durationStr) * 1000 : undefined;
+                        if (testResults.site.testing.test[i].status !== 'passed' && !havefailures) {
+                            const failureDurationStr = testResults.site.testing.test[i].measurements.get("Execution Time")?.value;
+                            const failureDuration = failureDurationStr ? parseFloat(failureDurationStr) * 1000 : undefined;
+                            const exitCode = testResults.site.testing.test[i].measurements.get("Exit Value")?.value;
+                            const completionStatus = testResults.site.testing.test[i].measurements.get("Completion Status")?.value;
 
-                        if (testResults.site.testing.test[0].status === 'passed') {
-                            run.passed(test, duration);
-                        } else {
-                            const exitCode = testResults.site.testing.test[0].measurements.get("Exit Value")?.value;
-                            const completionStatus = testResults.site.testing.test[0].measurements.get("Completion Status")?.value;
                             if (exitCode !== undefined) {
                                 this.ctestFailed(
                                     test,
                                     run,
-                                    new vscode.TestMessage(localize('test.failed.with.exit.code', 'Test failed with exit code {0}.', exitCode)),
-                                    duration
+                                    new vscode.TestMessage(localize('test.failed.with.exit.code', 'Test {0} failed with exit code {1}.', testName, exitCode)),
+                                    failureDuration
                                 );
                             } else if (completionStatus !== undefined) {
                                 this.ctestErrored(
                                     test,
                                     run,
-                                    new vscode.TestMessage(localize('test.failed.with.completion.status', 'Test failed with completion status "{0}".', completionStatus))
+                                    new vscode.TestMessage(localize('test.failed.with.completion.status', 'Test {0} failed with completion status "{1}".', testName, completionStatus))
                                 );
                             } else {
                                 this.ctestErrored(
                                     test,
                                     run,
-                                    new vscode.TestMessage(localize('test.failed', 'Test failed. Please check output for more information.'))
+                                    new vscode.TestMessage(localize('test.failed', 'Test {0} failed. Please check output for more information.', testName))
                                 );
                             }
+
+                            havefailures = true;
                             returnCode = -1;
                         }
-                    } else {
-                        this.ctestFailed(test, run, new vscode.TestMessage(localize('expect.one.test.results', 'Expecting one test result.')));
-                        returnCode = -1;
                     }
-                } else {
+                }
+
+                if (!foundTestResult && !havefailures) {
                     this.ctestFailed(test, run, new vscode.TestMessage(localize('test.results.not.found', 'Test results not found.')));
+                    havefailures = true;
                     returnCode = -1;
+                }
+
+                if (!havefailures) {
+                    run.passed(test, duration);
                 }
             }
         }
@@ -494,18 +518,14 @@ export class CTestDriver implements vscode.Disposable {
             return -2;
         }
 
-        const buildConfigArgs: string[] = [];
-        if (driver.useCMakePresets) {
-            const buildConfig = driver.testPreset?.configuration;
-            if (buildConfig) {
-                buildConfigArgs.push('-C', buildConfig);
-            }
-        } else {
-            buildConfigArgs.push('-C', driver.currentBuildType);
+        const ctestArgs = await this.getCTestArgs(driver, useCMakePresets);
+        if (!ctestArgs) {
+            log.info(localize('ctest.args.not.found', 'Could not get CTest arguments'));
+            return -3;
         }
         if (!driver.cmake.version || util.versionLess(driver.cmake.version, { major: 3, minor: 14, patch: 0 })) {
             // ctest --show-only=json-v1 was added in CMake 3.14
-            const result = await driver.executeCommand(ctestpath, ['-N', ...buildConfigArgs], undefined, { cwd: driver.binaryDir, silent: true }).result;
+            const result = await driver.executeCommand(ctestpath, ['-N', ...ctestArgs], undefined, { cwd: driver.binaryDir, silent: true }).result;
             if (result.retc !== 0) {
                 // There was an error running CTest. Odd...
                 log.error(localize('ctest.error', 'There was an error running ctest to determine available test executables'));
@@ -522,7 +542,7 @@ export class CTestDriver implements vscode.Disposable {
                 testExplorerRoot.children.add(initializedTestExplorer.createTestItem(test.name, test.name));
             }
         } else {
-            const result = await driver.executeCommand(ctestpath, ['--show-only=json-v1', ...buildConfigArgs], undefined, { cwd: driver.binaryDir, silent: true }).result;
+            const result = await driver.executeCommand(ctestpath, ['--show-only=json-v1', ...ctestArgs], undefined, { cwd: driver.binaryDir, silent: true }).result;
             if (result.retc !== 0) {
                 // There was an error running CTest. Odd...
                 log.error(localize('ctest.error', 'There was an error running ctest to determine available test executables'));
@@ -629,7 +649,8 @@ export class CTestDriver implements vscode.Disposable {
         const configs = launchConfig.get<vscode.DebugConfiguration[]>('configurations') ?? [];
         const workspaceConfigs = workspaceLaunchConfig?.get<vscode.DebugConfiguration[]>('configurations') ?? [];
         if (configs.length === 0 && workspaceConfigs.length === 0) {
-            throw(localize('no.launch.config', 'No launch configurations found.'));
+            log.error(localize('no.launch.config', 'No launch configurations found.'));
+            return;
         }
 
         interface ConfigItem extends vscode.QuickPickItem {